server/requirements.txt
# Core Django
Django==5.2.3

# Database
psycopg2-binary==2.9.9

# API Development
djangorestframework==3.16.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
python-jose==3.3.0

# Environment Variables
python-dotenv==1.0.1

# Image Processing
Pillow==10.2.0



# Testing
pytest==8.0.2
pytest-django==4.7.0

# Documentation
drf-yasg==1.21.7

# Security
django-environ==0.11.2

# Filtering
django-filter==24.3

# Async Support
channels==4.0.0
channels-redis==4.1.0

# Caching
redis==5.0.1

# Email
sendgrid==6.11.0
django-sendgrid-v5==1.2.1